Refactor part 2 - clean main.rs
